Track.vue --> now displays tracks
spotify.js --> fixed issue with token being reset from url

+++ src/store/modules/spotify.js
@@ -1,4 +1,4 @@
-import axios from "axios";
+import axios from 'axios';
 
 const state = {
   user: null,
@@ -12,45 +12,59 @@ const getters = {
 };
 
 const actions = {
-  setToken() {
-    const access_token = window.location.hash.slice(14, 300);
-    const refresh_token = window.location.href.slice(-131);
+  async setToken({ commit }) {
+    const tokenExists = await JSON.parse(localStorage.getItem('tokens'));
 
-    const tokens = {
-      access_token: access_token,
-      refresh_token: refresh_token
-    };
+    if (tokenExists) {
+      console.log('token already exists');
+    } else {
+      const access_token = window.location.hash.slice(14, 300);
+      const refresh_token = window.location.href.slice(-131);
 
-    localStorage.setItem("tokens", JSON.stringify(tokens));
+      const tokens = {
+        access_token: access_token,
+        refresh_token: refresh_token
+      };
+
+      localStorage.setItem('tokens', JSON.stringify(tokens));
+    }
   },
   async getUser({ commit }) {
-    const { access_token } = await JSON.parse(localStorage.getItem("tokens"));
+    const { access_token } = await JSON.parse(localStorage.getItem('tokens'));
 
-    const user = await axios.get("https://api.spotify.com/v1/me", {
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${access_token}`
-      }
-    });
-
-    commit("setUser", user.data);
-    commit("setUserLoggedIn", true);
-  },
-  async getUserLibraryTracks({ commit }) {
-    const { access_token } = await JSON.parse(localStorage.getItem("tokens"));
-
-    const tracks = await axios.get(
-      "https://api.spotify.com/v1/me/tracks?limit=50",
-      {
+    if (access_token) {
+      const user = await axios.get('https://api.spotify.com/v1/me', {
         headers: {
-          "Content-Type": "application/json",
+          'Content-Type': 'application/json',
           Authorization: `Bearer ${access_token}`
         }
-      }
-    );
+      });
+      commit('setUser', user.data);
+    }
+  },
+  async getUserLibraryTracks({ commit }) {
+    const { access_token } = await JSON.parse(localStorage.getItem('tokens'));
 
-    commit("setUserLibraryTracks", tracks.data.items);
-    console.log(tracks.data.items);
+    if (access_token) {
+      const tracks = await axios.get(
+        'https://api.spotify.com/v1/me/tracks?limit=10',
+        {
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${access_token}`
+          }
+        }
+      );
+
+      commit('setUserLibraryTracks', tracks.data.items);
+    }
+  },
+  async setUserLoggedInStatus({ commit }) {
+    const { access_token } = await JSON.parse(localStorage.getItem('tokens'));
+
+    if (access_token) commit('setUserLoggedIn', true);
+
+    console.log(state.userLoggedIn);
   }
 };
